+++ examples/cloudflare_worker/src/comlinks.ts
