+++ examples/node_js/index.mjs
@@ -1,5 +1,5 @@
 import { createServer } from 'http';
-import { SuperfaceClient } from '../../host/js/node/index.js';
+import { OneClient } from '../../host/js/node/index.js';
 
 async function startLocalhostServer() {
   const server = createServer((req, res) => {
@@ -21,7 +21,7 @@ async function startLocalhostServer() {
 async function main() {
   const server = await startLocalhostServer();
 
-  const client = new SuperfaceClient({ assetsPath: '../examples/comlinks/src' });
+  const client = new OneClient({ assetsPath: '../examples/comlinks/src' });
   const profile = await client.getProfile('wasm-sdk/example');
   const result = await profile
     .getUseCase('Example')
