+++ host/js/src/cloudflare/index.ts
@@ -254,7 +254,7 @@ class InternalClient {
   }
 }
 
-export class SuperfaceClient {
+export class OneClient {
   private internal: InternalClient;
 
   constructor(readonly options: ClientOptions = {}) {
